Move timing-utils from runtime/ sublayer to runtime-wrapper workspace src root
Previously: src/runtime-wrapper/src/runtime/timing-utils.ts
Now:        src/runtime-wrapper/src/timing-utils.ts

Rationale: getHighResolutionTime, getWallClockTime, and measureDuration are
generic timing helpers that are cross-cutting within the runtime-wrapper
workspace — they are consumed by both runtime/runtime-wrapper.ts and
websocket/client.ts (peer sublayers). Placing them inside runtime/ created a
false inter-layer dependency: websocket/ had to reach into runtime/ via
../runtime/timing-utils.js just to get a generic time function. Moving the
file to the workspace src root makes the shared-utility nature explicit and
removes the asymmetric coupling between peer sublayers.

All imports updated; public API (RuntimeWrapper.getHighResolutionTime etc.)
preserved through src/index.ts. All 282 runtime-wrapper tests continue to pass.
