+++ src/runtime-wrapper/src/index.ts
@@ -1,10 +1,14 @@
 import * as Runtime from "./runtime/index.js";
+import { getHighResolutionTime, getWallClockTime, measureDuration } from "./timing-utils.js";
 import * as Clients from "./websocket/index.js";
 
 // Export the RuntimeWrapper namespace as the primary public API
 export const RuntimeWrapper = Object.freeze({
     ...Runtime,
-    ...Clients
+    ...Clients,
+    getHighResolutionTime,
+    getWallClockTime,
+    measureDuration
 });
 
 export type {
@@ -71,3 +75,7 @@ export type {
     WebSocketPatchQueueManager
 } from "./websocket/index.js";
 export * as Clients from "./websocket/index.js";
+// Timing utilities are workspace-level cross-cutting helpers used by both the
+// runtime and websocket layers. They are exported directly from the workspace
+// root so consumers do not need to depend on either sublayer to access them.
+export { getHighResolutionTime, getWallClockTime, measureDuration } from "./timing-utils.js";
