+++ src/runtime-wrapper/src/timing-utils.ts
@@ -1,10 +1,14 @@
 /**
- * Timing utilities for the runtime wrapper.
+ * Shared timing utilities for the runtime-wrapper workspace.
  *
  * This module provides high-resolution timing for duration measurements using performance.now()
  * while preserving Date.now() for wall-clock timestamps. The performance API provides
  * sub-millisecond precision and monotonic timestamps that are immune to system clock adjustments,
  * making it ideal for measuring patch application durations in hot-reload scenarios.
+ *
+ * These helpers are cross-cutting within the workspace and are used by both the runtime
+ * and websocket layers, so they live at the workspace src root rather than inside either
+ * sublayer.
  */
 
 /**
