+++ src/runtime-wrapper/test/timing-utils.test.ts
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import { describe, it } from "node:test";
 
-import { getHighResolutionTime, getWallClockTime, measureDuration } from "../src/runtime/timing-utils.js";
+import { getHighResolutionTime, getWallClockTime, measureDuration } from "../src/timing-utils.js";
 
 await describe("timing utilities", async () => {
     await it("getWallClockTime returns a number", () => {
